Temporarily remove this module's tracking of php.ini.
This crazy thing is _way_ too complicated. With so many redirections
it's hard to figure out what's going where and why. I'll rewrite it
eventually, but for now, because of a collision that only _just_
appeared though it should have manifested months ago, I'll just remove
this module's tracking of php.ini.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>